Move all C source code under src/
- Also adjust includes and build to src/ layout
This was reported by @WojciechMula as a pet-pieve.
This makes 100% sense and bring some clarity to the repo organization
