--- allsources.c
@@ -1,6 +1,0 @@
-#include "src/custompickle/custompickle.c"
-#include "src/custompickle/pyhelpers.c"
-#include "src/custompickle/save/savebuffer.c"
-#include "src/custompickle/save/automaton_save.c"
-#include "src/custompickle/load/loadbuffer.c"
-#include "src/custompickle/load/module_automaton_load.c"
