+++ setup.cfg
@@ -1,5 +1,7 @@
 [metadata]
-license_file = LICENSE
+license_files =
+    LICENSE
+    README.rst
 
 [build_sphinx]
 source-dir = docs/
