+++ setup.py
@@ -39,27 +39,27 @@ else:
 module = Extension(
     'ahocorasick',
     sources=[
-        'pyahocorasick.c',
+        'src/pyahocorasick.c',
     ],
     define_macros=macros,
     depends=[
-        'common.h',
-        'Automaton.c',
-        'Automaton.h',
-        'Automaton_pickle.c',
-        'AutomatonItemsIter.c',
-        'AutomatonItemsIter.h',
-        'AutomatonSearchIter.c',
-        'AutomatonSearchIter.h',
-        'AutomatonSearchIterLong.c',
-        'AutomatonSearchIterLong.h',
-        'trie.c',
-        'trie.h',
-        'slist.c',
-        'utils.c',
-        'trienode.c',
-        'trienode.h',
-        'msinttypes/stdint.h',
+        'src/common.h',
+        'src/Automaton.c',
+        'src/Automaton.h',
+        'src/Automaton_pickle.c',
+        'src/AutomatonItemsIter.c',
+        'src/AutomatonItemsIter.h',
+        'src/AutomatonSearchIter.c',
+        'src/AutomatonSearchIter.h',
+        'src/AutomatonSearchIterLong.c',
+        'src/AutomatonSearchIterLong.h',
+        'src/trie.c',
+        'src/trie.h',
+        'src/slist.c',
+        'src/utils.c',
+        'src/trienode.c',
+        'src/trienode.h',
+        'src/msinttypes/stdint.h',
         'src/inline_doc.h',
         'src/pickle/pickle.h',
         'src/pickle/pickle_data.h',
