+++ src/Automaton.h
