+++ src/AutomatonItemsIter.c
