+++ src/AutomatonItemsIter.h
