+++ src/AutomatonSearchIter.c
