+++ src/AutomatonSearchIter.h
