+++ src/custompickle/load/module_automaton_load.c
@@ -1,6 +1,6 @@
 #include "module_automaton_load.h"
 
-#include "../../../Automaton.h"
+#include "../../Automaton.h"
 #include "loadbuffer.h"
