+++ src/custompickle/save/automaton_save.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../../common.h"
+#include "../../common.h"
 
 PyObject*
 automaton_save(PyObject* self, PyObject* args);
