+++ src/slist.h
