+++ src/trie.h
