+++ src/trienode.c
